pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "device_control"
description = "SCExAO device control"
readme = "README.md"
authors = [
    {name = "Vincent Deo", email = "vdeo@naoj.org"},
    {name = "Miles Lucas", email = "mdlucas@hawaii.edu"}
]
dependencies = [
    "docopt==0.6.*",
    "libtmux==0.8.*",
    "pyserial>=3,<4",
    "tomli",
    "tomli-w",
    "zaber-motion>=2,<3",
    "tqdm>=4,<5",
]
dynamic = ["version"]

[project.optional-dependencies]
test = [
    "pytest==7.*",
]

[project.scripts]
# daemons
vampires_daemon = "device_control.daemons.vampires:main"
# VAMPIRES devices
vampires_diffwheel = "device_control.vampires.vampires_diffwheel:main"
vampires_beamsplitter = "device_control.vampires.vampires_beamsplitter:main"
vampires_focus = "device_control.vampires.vampires_focus:main"
vampires_camfocus = "device_control.vampires.vampires_camfocus:main"
vampires_fieldstop = "device_control.vampires.vampires_fieldstop:main"
vampires_filter = "device_control.vampires.vampires_filter:main"
vampires_pupil = "device_control.vampires.vampires_pupil:main"
vampires_qwp = "device_control.vampires.vampires_qwp:main"

[tool.setuptools.dynamic]
version = {attr = "device_control.__version__"}

[tool.coverage.run]
source = ["device_control"]
